minor reorg per susan PR review
